Rename: VoucherManager 패키지 변경
voucher -> controller
